+++ data_service/service/models/home.py
@@ -7,7 +7,7 @@ class Home(models.Model):
     date = models.DateTimeField("date")
     price = models.FloatField()
     bedrooms = models.IntegerField()
-    bathrooms = models.IntegerField()
+    bathrooms = models.FloatField()
     sqft_living = models.FloatField()
     sqft_lot = models.FloatField()
     floors = models.IntegerField()
